Fix: remove append_value for PYTHON
Doing this breaks, if multiple things edit the PYTHON env, because PYTHON
will be set to a list like `["python3", "python3"]` then, which obviously
is not a valid command to run python.

Currently this actively breaks `chip-io`.

Finally setting this is no longer necessary, as the ASIC container
defaults to python 3.

Change-Id: I8b8b932d5d5dc47eae22ee371eddeb584e6f8013